Refactor settings page, make server component and add SettingsView
